Use json.RawMessage for GetArchiveRequest
For some instances, unmarshaling JSON fails when a property
is set as `[]byte` and the raw value is base64 encoded.

To fix it, since we don't need to decode the `GetArchiveRequest`
when we unmarshal the JSON, set it to `json.RawMessage` instead of
`[]byte`.

See https://golang.org/pkg/encoding/json/#RawMessage for more
info about `json.RawMessage`.