Allow Content-Type from Rails
Now, when Rails sets a Content-Type, Workhorse
will compare it with the one it detectes.

If it's too different, ie application vs text,
Workhorse will return the detected one, otherwise
it will return the existing one.

We do this because the content type detected by
Rails is more accurate than the one detected by
Workhorse.